add custom ObjectIdDeserializer to use String instead of ObjectId as type
When using ObjectId, the response is encoded with {timestamp, date} which is not what I want, but more importantly, it seem to create an issue / bug where if it runs through the Converter class, the ObjectId will always be different from the original, thus making the conversion between mongo > es having issue. For example, mongo created id is 1, but when index into es, it will be id 2.